requirements.txt
chromadb
requests
beautifulsoup4
openai
python-dotenv